+++ executors/docker/docker_command_integration_test.go
@@ -2255,6 +2255,50 @@ func Test_FF_USE_INIT_WITH_DOCKER_EXECUTOR(t *testing.T) {
 	}
 }
 
+func Test_FF_USE_INIT_WITH_DOCKER_EXECUTOR(t *testing.T) {
+	test.SkipIfGitLabCIOn(t, test.OSWindows)
+	helpers.SkipIntegrationTests(t, "docker", "info")
+
+	tests := map[string]bool{
+		"use init":        true,
+		"do not use init": false,
+	}
+
+	for name, useInit := range tests {
+		t.Run(name, func(t *testing.T) {
+			successfulBuild, err := common.GetRemoteBuildResponse("ps -A")
+			assert.NoError(t, err)
+			successfulBuild.Image.Name = common.TestAlpineImage
+
+			build := &common.Build{
+				JobResponse: successfulBuild,
+				Runner: &common.RunnerConfig{
+					RunnerSettings: common.RunnerSettings{
+						Executor: "docker",
+						Docker:   &common.DockerConfig{},
+					},
+				},
+			}
+
+			if useInit {
+				build.Variables = append(build.Variables, common.JobVariable{
+					Key:   "FF_USE_INIT_WITH_DOCKER_EXECUTOR",
+					Value: "true",
+				})
+			}
+
+			out := bytes.NewBuffer(nil)
+			assert.NoError(t, build.Run(&common.Config{}, &common.Trace{Writer: out}))
+
+			if useInit {
+				assert.Regexp(t, "1 root      0:00 /sbin/docker-init -- sh", out.String())
+			} else {
+				assert.NotRegexp(t, "1 root      0:00 /sbin/docker-init -- sh", out.String())
+			}
+		})
+	}
+}
+
 func TestDockerCommandWithPlatform(t *testing.T) {
 	test.SkipIfGitLabCIOn(t, test.OSWindows)
 	helpers.SkipIntegrationTests(t, "docker", "info")